refactor: refactor the basic work and initialization of badge on v2
